+++ packages/liveblocks-react-ui/.eslintrc.js
@@ -26,34 +26,6 @@ module.exports = {
           "ImportDeclaration[source.value='react'] ImportSpecifier[imported.name='use']",
         message: "use is only available on React >=19.",
       },
-      {
-        selector:
-          "ImportDeclaration[source.value='react'] ImportSpecifier[imported.name='useSyncExternalStore']",
-        message:
-          "useSyncExternalStore is only available on React >=18. Import it from 'use-sync-external-store/shim/index.js' instead.",
-      },
-      {
-        selector:
-          "ImportDeclaration[source.value='react'] ImportSpecifier[imported.name='useId']",
-        message:
-          "useId is only available on React >=18. Import it from '/src/utils/use-id' instead.",
-      },
-      {
-        selector:
-          "ImportDeclaration[source.value='react'] ImportSpecifier[imported.name='useTransition']",
-        message:
-          "useTransition is only available on React >=18. Import it from '/src/utils/use-transition' instead.",
-      },
-      {
-        selector:
-          "ImportDeclaration[source.value='react'] ImportSpecifier[imported.name='useDeferredValue']",
-        message: "useDeferredValue is only available on React >=18.",
-      },
-      {
-        selector:
-          "ImportDeclaration[source.value='react'] ImportSpecifier[imported.name='useInsertionEffect']",
-        message: "useInsertionEffect is only available on React >=18.",
-      },
     ],
 
     // ----------------------------------------------------------------------
@@ -92,6 +64,12 @@ module.exports = {
         "@typescript-eslint/no-unsafe-return": "off",
         "@typescript-eslint/unbound-method": "off",
         "@typescript-eslint/no-floating-promises": "off",
+      },
+    },
+    {
+      files: ["rollup.config.mjs"],
+      env: {
+        node: true,
       },
     },
   ],