+++ packages/liveblocks-react-ui/src/components/internal/User.tsx
@@ -2,7 +2,7 @@
 
 import { useUser } from "@liveblocks/react";
 import type { ComponentProps } from "react";
-import React, { useMemo } from "react";
+import { useMemo } from "react";
 
 import { useOverrides } from "../../overrides";
 import { useCurrentUserId } from "../../shared";