+++ packages/liveblocks-react/src/__tests__/useInboxNotifications.test.tsx
@@ -10,7 +10,7 @@ import {
   waitFor,
 } from "@testing-library/react";
 import { setupServer } from "msw/node";
-import React, { Suspense } from "react";
+import { Suspense } from "react";
 import { ErrorBoundary, type FallbackProps } from "react-error-boundary";
 
 import { dummyThreadData, dummyThreadInboxNotificationData } from "./_dummies";
@@ -284,15 +284,14 @@ describe("useInboxNotifications", () => {
       umbrellaStore,
     } = createContextsForTest();
 
-    umbrellaStore.baseThreadsDB.upsert(thread1);
-    umbrellaStore.baseThreadsDB.upsert(thread2);
+    umbrellaStore.threads.upsert(thread1);
+    umbrellaStore.threads.upsert(thread2);
 
-    umbrellaStore.force_set_notifications((prev) => ({
-      ...prev,
+    umbrellaStore.force_set_notifications((lut) => {
       // Explicitly set the order to be reversed to test that the hook sorts the notifications
-      [oldInboxNotification.id]: oldInboxNotification,
-      [newInboxNotification.id]: newInboxNotification,
-    }));
+      lut.set(oldInboxNotification.id, oldInboxNotification);
+      lut.set(newInboxNotification.id, newInboxNotification);
+    });
 
     const { result, unmount } = renderHook(() => useInboxNotifications(), {
       wrapper: ({ children }) => (