+++ packages/liveblocks-react/src/__tests__/useThreads.test.tsx
@@ -25,7 +25,7 @@ import {
 } from "msw";
 import { setupServer } from "msw/node";
 import type { ReactNode } from "react";
-import React, { Suspense } from "react";
+import { createContext, Suspense, useContext, useState } from "react";
 import { ErrorBoundary } from "react-error-boundary";
 
 import { dummyThreadData, dummyThreadInboxNotificationData } from "./_dummies";
@@ -1071,12 +1071,12 @@ describe("useThreads", () => {
       room: { RoomProvider, useThreads },
     } = createContextsForTest();
 
-    const RoomIdDispatchContext = React.createContext<
+    const RoomIdDispatchContext = createContext<
       ((value: string) => void) | null
     >(null);
 
     const Wrapper = ({ children }: { children: ReactNode }) => {
-      const [roomId, setRoomId] = React.useState(room1Id);
+      const [roomId, setRoomId] = useState(room1Id);
 
       return (
         <RoomIdDispatchContext.Provider value={setRoomId}>
@@ -1086,7 +1086,7 @@ describe("useThreads", () => {
     };
 
     const useThreadsContainer = () => {
-      const setRoomId = React.useContext(RoomIdDispatchContext);
+      const setRoomId = useContext(RoomIdDispatchContext);
       const state = useThreads();
       return { state, setRoomId };
     };
@@ -1453,10 +1453,10 @@ describe("useThreads", () => {
       umbrellaStore,
     } = createContextsForTest();
 
-    const db = umbrellaStore.baseThreadsDB;
+    const db = umbrellaStore.threads;
     db.upsert(thread1);
     db.upsert(thread2WithDeletedAt);
-    umbrellaStore.force_set_versions((state) => ({ ...state }));
+    umbrellaStore.invalidateEntireStore();
 
     const { result, unmount } = renderHook(
       () => useThreads({ query: { metadata: {} } }),