+++ packages/liveblocks-react/src/room.tsx
@@ -43,7 +43,6 @@ import {
   console,
   createCommentId,
   createThreadId,
-  deprecateIf,
   errorIf,
   HttpError,
   kInternal,
@@ -51,8 +50,16 @@ import {
   makePoller,
   ServerMsgCode,
 } from "@liveblocks/core";
-import * as React from "react";
-import { useSyncExternalStoreWithSelector } from "use-sync-external-store/shim/with-selector.js";
+import type { Context } from "react";
+import {
+  useCallback,
+  useEffect,
+  useMemo,
+  useRef,
+  useState,
+  useSyncExternalStore,
+  version as reactVersion,
+} from "react";
 
 import { config } from "./config";
 import { RoomContext, useIsInsideRoom, useRoomOrNull } from "./contexts";
@@ -110,47 +117,21 @@ import {
 import type { UmbrellaStore } from "./umbrella-store";
 import { useScrollToCommentOnLoadEffect } from "./use-scroll-to-comment-on-load-effect";
 import { useSignal } from "./use-signal";
+import { useSyncExternalStoreWithSelector } from "./use-sync-external-store-with-selector";
 
 const noop = () => {};
 const identity: <T>(x: T) => T = (x) => x;
 
-const missing_unstable_batchedUpdates = (
-  reactVersion: number,
-  roomId: string
-) =>
-  `We noticed you’re using React ${reactVersion}. Please pass unstable_batchedUpdates at the RoomProvider level until you’re ready to upgrade to React 18:
-
-    import { unstable_batchedUpdates } from "react-dom";  // or "react-native"
-
-    <RoomProvider id=${JSON.stringify(
-      roomId
-    )} ... unstable_batchedUpdates={unstable_batchedUpdates}>
-      ...
-    </RoomProvider>
-
-Why? Please see https://liveblocks.io/docs/platform/troubleshooting#stale-props-zombie-child for more information`;
-
-const superfluous_unstable_batchedUpdates =
-  "You don’t need to pass unstable_batchedUpdates to RoomProvider anymore, since you’re on React 18+ already.";
-
-function useSyncExternalStore<Snapshot>(
-  s: (onStoreChange: () => void) => () => void,
-  gs: () => Snapshot,
-  gss: undefined | null | (() => Snapshot)
-): Snapshot {
-  return useSyncExternalStoreWithSelector(s, gs, gss, identity);
-}
-
 const STABLE_EMPTY_LIST = Object.freeze([]);
 
 // Don't try to inline this. This function is intended to be a stable
-// reference, to avoid a React.useCallback() wrapper.
+// reference, to avoid a useCallback() wrapper.
 function alwaysEmptyList() {
   return STABLE_EMPTY_LIST;
 }
 
 // Don't try to inline this. This function is intended to be a stable
-// reference, to avoid a React.useCallback() wrapper.
+// reference, to avoid a useCallback() wrapper.
 function alwaysNull() {
   return null;
 }
@@ -271,10 +252,10 @@ function makeRoomExtrasForClient<M extends BaseMetadata>(client: OpaqueClient) {
 
   function onMutationFailure(
     innerError: Error,
-    optimisticUpdateId: string,
+    optimisticId: string,
     createPublicError: (error: Error) => CommentsError<M>
   ) {
-    store.removeOptimisticUpdate(optimisticUpdateId);
+    store.optimisticUpdates.remove(optimisticId);
 
     if (innerError instanceof HttpError) {
       const error = handleApiError(innerError);
@@ -413,7 +394,7 @@ function makeRoomContextBundle<
   const shared = createSharedContext<U>(client);
 
   const bundle: RoomContextBundle<P, S, U, E, M> = {
-    RoomContext: RoomContext as React.Context<TRoom | null>,
+    RoomContext: RoomContext as Context<TRoom | null>,
     RoomProvider: RoomProvider_withImplicitLiveblocksProvider,
 
     useRoom,
@@ -471,7 +452,7 @@ function makeRoomContextBundle<
     ...shared.classic,
 
     suspense: {
-      RoomContext: RoomContext as React.Context<TRoom | null>,
+      RoomContext: RoomContext as Context<TRoom | null>,
       RoomProvider: RoomProvider_withImplicitLiveblocksProvider,
 
       useRoom,
@@ -551,36 +532,35 @@ function RoomProvider<
   M extends BaseMetadata,
 >(props: RoomProviderProps<P, S>) {
   const client = useClient<U>();
-  const [cache] = React.useState(
+  const [cache] = useState(
     () => new Map<string, RoomLeavePair<P, S, U, E, M>>()
   );
 
   // Produce a version of client.enterRoom() that when called for the same
   // room ID multiple times, will not keep producing multiple leave
   // functions, but instead return the cached one.
-  const stableEnterRoom: typeof client.enterRoom<P, S, E, M> =
-    React.useCallback(
-      (
-        roomId: string,
-        options: EnterOptions<P, S>
-      ): RoomLeavePair<P, S, U, E, M> => {
-        const cached = cache.get(roomId);
-        if (cached) return cached;
+  const stableEnterRoom: typeof client.enterRoom<P, S, E, M> = useCallback(
+    (
+      roomId: string,
+      options: EnterOptions<P, S>
+    ): RoomLeavePair<P, S, U, E, M> => {
+      const cached = cache.get(roomId);
+      if (cached) return cached;
 
-        const rv = client.enterRoom<P, S, E, M>(roomId, options);
+      const rv = client.enterRoom<P, S, E, M>(roomId, options);
 
-        // Wrap the leave function to also delete the cached value
-        const origLeave = rv.leave;
-        rv.leave = () => {
-          origLeave();
-          cache.delete(roomId);
-        };
+      // Wrap the leave function to also delete the cached value
+      const origLeave = rv.leave;
+      rv.leave = () => {
+        origLeave();
+        cache.delete(roomId);
+      };
 
-        cache.set(roomId, rv);
-        return rv;
-      },
-      [client, cache]
-    );
+      cache.set(roomId, rv);
+      return rv;
+    },
+    [client, cache]
+  );
 
   //
   // RATIONALE:
@@ -646,15 +626,11 @@ function RoomProviderInner<
       throw new Error("RoomProvider id property should be a string.");
     }
 
-    const majorReactVersion = parseInt(React.version) || 1;
-    const oldReactVersion = majorReactVersion < 18;
+    const majorReactVersion = parseInt(reactVersion) || 1;
+    const requiredVersion = 18;
     errorIf(
-      oldReactVersion && props.unstable_batchedUpdates === undefined,
-      missing_unstable_batchedUpdates(majorReactVersion, roomId)
-    );
-    deprecateIf(
-      !oldReactVersion && props.unstable_batchedUpdates !== undefined,
-      superfluous_unstable_batchedUpdates
+      majorReactVersion < requiredVersion,
+      `React ${requiredVersion} or higher is required (you’re on ${reactVersion})`
     );
   }
 
@@ -663,18 +639,17 @@ function RoomProviderInner<
   const frozenProps = useInitial({
     initialPresence: props.initialPresence,
     initialStorage: props.initialStorage,
-    unstable_batchedUpdates: props.unstable_batchedUpdates,
     autoConnect: props.autoConnect ?? typeof window !== "undefined",
   }) as EnterOptions<P, S>;
 
-  const [{ room }, setRoomLeavePair] = React.useState(() =>
+  const [{ room }, setRoomLeavePair] = useState(() =>
     stableEnterRoom(roomId, {
       ...frozenProps,
       autoConnect: false, // Deliberately using false here on the first render, see below
     })
   );
 
-  React.useEffect(() => {
+  useEffect(() => {
     const { store } = getRoomExtrasForClient(client);
 
     async function handleCommentEvent(message: CommentsEventServerMsg) {
@@ -693,7 +668,7 @@ function RoomProviderInner<
         store.deleteThread(message.threadId, null);
         return;
       }
-      const { thread, inboxNotification } = info;
+      const { thread, inboxNotification: maybeNotification } = info;
 
       const existingThread = store
         .get1_threads()
@@ -709,10 +684,17 @@ function RoomProviderInner<
           // If the thread doesn't exist in the local cache, we do not update it with the server data as an optimistic update could have deleted the thread locally.
           if (!existingThread) break;
 
-          store.updateThreadAndNotification(thread, inboxNotification);
+          store.updateThreadifications(
+            [thread],
+            maybeNotification ? [maybeNotification] : []
+          );
           break;
+
         case ServerMsgCode.COMMENT_CREATED:
-          store.updateThreadAndNotification(thread, inboxNotification);
+          store.updateThreadifications(
+            [thread],
+            maybeNotification ? [maybeNotification] : []
+          );
           break;
         default:
           break;
@@ -724,7 +706,7 @@ function RoomProviderInner<
     );
   }, [client, room]);
 
-  React.useEffect(() => {
+  useEffect(() => {
     const pair = stableEnterRoom(roomId, frozenProps);
 
     setRoomLeavePair(pair);
@@ -778,10 +760,10 @@ function useStatus(): Status {
 
 /** @private - Internal API, do not rely on it. */
 function useReportTextEditor(editor: TextEditorType, rootKey: string): void {
-  const isReported = React.useRef<boolean>(false);
+  const isReported = useRef<boolean>(false);
   const room = useRoom();
 
-  React.useEffect(() => {
+  useEffect(() => {
     // We use a "locker" reference to avoid to spam / harass our backend
     // and to not add / remove subscribers in case when the text editor type
     // has been already reported.
@@ -805,14 +787,14 @@ function useReportTextEditor(editor: TextEditorType, rootKey: string): void {
 function useYjsProvider(): IYjsProvider | undefined {
   const room = useRoom();
 
-  const subscribe = React.useCallback(
+  const subscribe = useCallback(
     (onStoreChange: () => void): UnsubscribeCallback => {
       return room[kInternal].yjsProviderDidChange.subscribe(onStoreChange);
     },
     [room]
   );
 
-  const getSnapshot = React.useCallback((): IYjsProvider | undefined => {
+  const getSnapshot = useCallback((): IYjsProvider | undefined => {
     return room[kInternal].getYjsProvider();
   }, [room]);
 
@@ -822,7 +804,7 @@ function useYjsProvider(): IYjsProvider | undefined {
 /** @private - Internal API, do not rely on it. */
 function useCreateTextMention(): (userId: string, mentionId: string) => void {
   const room = useRoom();
-  return React.useCallback(
+  return useCallback(
     (userId: string, mentionId: string): void => {
       room[kInternal]
         .createTextMention(userId, mentionId)
@@ -840,7 +822,7 @@ function useCreateTextMention(): (userId: string, mentionId: string) => void {
 /** @private - Internal API, do not rely on it. */
 function useDeleteTextMention(): (mentionId: string) => void {
   const room = useRoom();
-  return React.useCallback(
+  return useCallback(
     (mentionId: string): void => {
       room[kInternal].deleteTextMention(mentionId).catch((err): void => {
         console.error(`Cannot delete text mention '${mentionId}'`, err);
@@ -894,10 +876,10 @@ function useStorageStatusImmediate(): StorageStatus {
 
 function useStorageStatusSmooth(): StorageStatus {
   const room = useRoom();
-  const [status, setStatus] = React.useState(room.getStorageStatus);
+  const [status, setStatus] = useState(room.getStorageStatus);
   const oldStatus = useLatest(room.getStorageStatus());
 
-  React.useEffect(() => {
+  useEffect(() => {
     let timeoutId: ReturnType<typeof setTimeout>;
     const unsub = room.events.storageStatus.subscribe((newStatus) => {
       if (
@@ -940,7 +922,7 @@ function useBroadcastEvent<E extends Json>(): (
   options?: BroadcastOptions
 ) => void {
   const room = useRoom<never, never, never, E, never>();
-  return React.useCallback(
+  return useCallback(
     (
       event: E,
       options: BroadcastOptions = { shouldQueueEventIfNotReady: false }
@@ -956,7 +938,7 @@ function useOthersListener<P extends JsonObject, U extends BaseUserMeta>(
 ) {
   const room = useRoom<P, never, U, never, never>();
   const savedCallback = useLatest(callback);
-  React.useEffect(
+  useEffect(
     () => room.events.others.subscribe((event) => savedCallback.current(event)),
     [room, savedCallback]
   );
@@ -987,7 +969,7 @@ function useLostConnectionListener(
 ): void {
   const room = useRoom();
   const savedCallback = useLatest(callback);
-  React.useEffect(
+  useEffect(
     () =>
       room.events.lostConnection.subscribe((event) =>
         savedCallback.current(event)
@@ -1008,7 +990,7 @@ function useLostConnectionListener(
 function useErrorListener(callback: (err: LiveblocksError) => void): void {
   const room = useRoom();
   const savedCallback = useLatest(callback);
-  React.useEffect(
+  useEffect(
     () => room.events.error.subscribe((e) => savedCallback.current(e)),
     [room, savedCallback]
   );
@@ -1021,7 +1003,7 @@ function useEventListener<
 >(callback: (data: RoomEventMessage<P, U, E>) => void): void {
   const room = useRoom<P, never, U, E, never>();
   const savedCallback = useLatest(callback);
-  React.useEffect(() => {
+  useEffect(() => {
     const listener = (eventData: RoomEventMessage<P, U, E>) => {
       savedCallback.current(eventData);
     };
@@ -1093,7 +1075,7 @@ function useSelf<P extends JsonObject, U extends BaseUserMeta, T>(
   const getSnapshot: () => Snapshot = room.getSelf;
 
   const selector = maybeSelector ?? (identity as (me: User<P, U>) => T);
-  const wrappedSelector = React.useCallback(
+  const wrappedSelector = useCallback(
     (me: Snapshot): Selection => (me !== null ? selector(me) : null),
     [selector]
   );
@@ -1157,13 +1139,13 @@ function useOthersMapped<P extends JsonObject, U extends BaseUserMeta, T>(
   itemSelector: (other: User<P, U>) => T,
   itemIsEqual?: (prev: T, curr: T) => boolean
 ): ReadonlyArray<readonly [connectionId: number, data: T]> {
-  const wrappedSelector = React.useCallback(
+  const wrappedSelector = useCallback(
     (others: readonly User<P, U>[]) =>
       others.map((other) => [other.connectionId, itemSelector(other)] as const),
     [itemSelector]
   );
 
-  const wrappedIsEqual = React.useCallback(
+  const wrappedIsEqual = useCallback(
     (
       a: ReadonlyArray<readonly [connectionId: number, data: T]>,
       b: ReadonlyArray<readonly [connectionId: number, data: T]>
@@ -1211,7 +1193,7 @@ function useOther<P extends JsonObject, U extends BaseUserMeta, T>(
   selector: (other: User<P, U>) => T,
   isEqual?: (prev: T, curr: T) => boolean
 ): T {
-  const wrappedSelector = React.useCallback(
+  const wrappedSelector = useCallback(
     (others: readonly User<P, U>[]) => {
       // TODO: Make this O(1) instead of O(n)?
       const other = others.find((other) => other.connectionId === connectionId);
@@ -1220,7 +1202,7 @@ function useOther<P extends JsonObject, U extends BaseUserMeta, T>(
     [connectionId, selector]
   );
 
-  const wrappedIsEqual = React.useCallback(
+  const wrappedIsEqual = useCallback(
     (prev: T | NotFound, curr: T | NotFound): boolean => {
       if (prev === NOT_FOUND || curr === NOT_FOUND) {
         return prev === curr;
@@ -1266,13 +1248,13 @@ function useStorage<S extends LsonObject, T>(
   const room = useRoom<never, S, never, never, never>();
   const rootOrNull = useMutableStorageRoot<S>();
 
-  const wrappedSelector = React.useCallback(
+  const wrappedSelector = useCallback(
     (rootOrNull: Snapshot): Selection =>
       rootOrNull !== null ? selector(rootOrNull) : null,
     [selector]
   );
 
-  const subscribe = React.useCallback(
+  const subscribe = useCallback(
     (onStoreChange: () => void) =>
       rootOrNull !== null
         ? room.subscribe(rootOrNull, onStoreChange, { isDeep: true })
@@ -1280,7 +1262,7 @@ function useStorage<S extends LsonObject, T>(
     [room, rootOrNull]
   );
 
-  const getSnapshot = React.useCallback((): Snapshot => {
+  const getSnapshot = useCallback((): Snapshot => {
     if (rootOrNull === null) {
       return null;
     } else {
@@ -1310,7 +1292,7 @@ function useMutation<
   F extends (context: MutationContext<P, S, U>, ...args: any[]) => any,
 >(callback: F, deps: readonly unknown[]): OmitFirstArg<F> {
   const room = useRoom<P, S, U, E, M>();
-  return React.useMemo(
+  return useMemo(
     () => {
       return ((...args) =>
         // eslint-disable-next-line @typescript-eslint/no-unsafe-return
@@ -1344,7 +1326,7 @@ function useThreads<M extends BaseMetadata>(
 
   const poller = getOrCreateThreadsPollerForRoomId(room.id);
 
-  React.useEffect(
+  useEffect(
     () => {
       void store.waitUntilRoomThreadsLoaded(room.id, options.query);
     }
@@ -1358,19 +1340,19 @@ function useThreads<M extends BaseMetadata>(
     //    *next* render after that, a *new* fetch/promise will get created.
   );
 
-  React.useEffect(() => {
+  useEffect(() => {
     poller.inc();
     poller.pollNowIfStale();
     return () => poller.dec();
   }, [poller]);
 
-  // TODO(vincent+nimesh) There is a disconnect between this getter and subscriber! It's unclear
+  // XXX_vincent There is a disconnect between this getter and subscriber! It's unclear
   // why the getRoomThreadsLoadingState getter should be paired with subscribe1
   // and not subscribe2 from the outside! (The reason is that
   // getRoomThreadsLoadingState internally uses `get1` not `get2`.) This is
   // strong evidence that getRoomThreadsLoadingState itself wants to be
   // a Signal! Once we make it a Signal, we can simply use `useSignal()` here! ❤️
-  const getter = React.useCallback(
+  const getter = useCallback(
     () => store.getRoomThreadsLoadingState(room.id, options.query),
     [store, room.id, options.query]
   );
@@ -1398,7 +1380,7 @@ function useCommentsErrorListener<M extends BaseMetadata>(
   const savedCallback = useLatest(callback);
   const { commentsErrorEventSource } = getRoomExtrasForClient<M>(client);
 
-  React.useEffect(() => {
+  useEffect(() => {
     return commentsErrorEventSource.subscribe(savedCallback.current);
   }, [savedCallback, commentsErrorEventSource]);
 }
@@ -1417,7 +1399,7 @@ function useCreateRoomThread<M extends BaseMetadata>(
 ): (options: CreateThreadOptions<M>) => ThreadData<M> {
   const client = useClient();
 
-  return React.useCallback(
+  return useCallback(
     (options: CreateThreadOptions<M>): ThreadData<M> => {
       const body = options.body;
       const metadata = options.metadata ?? ({} as M);
@@ -1450,7 +1432,7 @@ function useCreateRoomThread<M extends BaseMetadata>(
       };
 
       const { store, onMutationFailure } = getRoomExtrasForClient(client);
-      const optimisticUpdateId = store.addOptimisticUpdate({
+      const optimisticId = store.optimisticUpdates.add({
         type: "create-thread",
         thread: newThread,
         roomId,
@@ -1470,12 +1452,12 @@ function useCreateRoomThread<M extends BaseMetadata>(
         .then(
           (thread) => {
             // Replace the optimistic update by the real thing
-            store.createThread(optimisticUpdateId, thread);
+            store.createThread(optimisticId, thread);
           },
           (err: Error) =>
             onMutationFailure(
               err,
-              optimisticUpdateId,
+              optimisticId,
               (err) =>
                 new CreateThreadError(err, {
                   roomId,
@@ -1499,7 +1481,7 @@ function useDeleteThread(): (threadId: string) => void {
 
 function useDeleteRoomThread(roomId: string): (threadId: string) => void {
   const client = useClient();
-  return React.useCallback(
+  return useCallback(
     (threadId: string): void => {
       const { store, onMutationFailure } = getRoomExtrasForClient(client);
 
@@ -1510,7 +1492,7 @@ function useDeleteRoomThread(roomId: string): (threadId: string) => void {
         throw new Error("Only the thread creator can delete the thread");
       }
 
-      const optimisticUpdateId = store.addOptimisticUpdate({
+      const optimisticId = store.optimisticUpdates.add({
         type: "delete-thread",
         roomId,
         threadId,
@@ -1520,12 +1502,12 @@ function useDeleteRoomThread(roomId: string): (threadId: string) => void {
       client[kInternal].httpClient.deleteThread({ roomId, threadId }).then(
         () => {
           // Replace the optimistic update by the real thing
-          store.deleteThread(threadId, optimisticUpdateId);
+          store.deleteThread(threadId, optimisticId);
         },
         (err: Error) =>
           onMutationFailure(
             err,
-            optimisticUpdateId,
+            optimisticId,
             (err) => new DeleteThreadError(err, { roomId, threadId })
           )
       );
@@ -1540,7 +1522,7 @@ function useEditThreadMetadata<M extends BaseMetadata>() {
 
 function useEditRoomThreadMetadata<M extends BaseMetadata>(roomId: string) {
   const client = useClient();
-  return React.useCallback(
+  return useCallback(
     (options: EditThreadMetadataOptions<M>): void => {
       if (!options.metadata) {
         return;
@@ -1551,7 +1533,7 @@ function useEditRoomThreadMetadata<M extends BaseMetadata>(roomId: string) {
       const updatedAt = new Date();
 
       const { store, onMutationFailure } = getRoomExtrasForClient(client);
-      const optimisticUpdateId = store.addOptimisticUpdate({
+      const optimisticId = store.optimisticUpdates.add({
         type: "edit-thread-metadata",
         metadata,
         threadId,
@@ -1563,16 +1545,11 @@ function useEditRoomThreadMetadata<M extends BaseMetadata>(roomId: string) {
         .then(
           (metadata) =>
             // Replace the optimistic update by the real thing
-            store.patchThread(
-              threadId,
-              optimisticUpdateId,
-              { metadata },
-              updatedAt
-            ),
+            store.patchThread(threadId, optimisticId, { metadata }, updatedAt),
           (err: Error) =>
             onMutationFailure(
               err,
-              optimisticUpdateId,
+              optimisticId,
               (error) =>
                 new EditThreadMetadataError(error, {
                   roomId,
@@ -1604,7 +1581,7 @@ function useCreateRoomComment(
   roomId: string
 ): (options: CreateCommentOptions) => CommentData {
   const client = useClient();
-  return React.useCallback(
+  return useCallback(
     ({ threadId, body, attachments }: CreateCommentOptions): CommentData => {
       const commentId = createCommentId();
       const createdAt = new Date();
@@ -1622,7 +1599,7 @@ function useCreateRoomComment(
       };
 
       const { store, onMutationFailure } = getRoomExtrasForClient(client);
-      const optimisticUpdateId = store.addOptimisticUpdate({
+      const optimisticId = store.optimisticUpdates.add({
         type: "create-comment",
         comment,
       });
@@ -1634,12 +1611,12 @@ function useCreateRoomComment(
         .then(
           (newComment) => {
             // Replace the optimistic update by the real thing
-            store.createComment(newComment, optimisticUpdateId);
+            store.createComment(newComment, optimisticId);
           },
           (err: Error) =>
             onMutationFailure(
               err,
-              optimisticUpdateId,
+              optimisticId,
               (err) =>
                 new CreateCommentError(err, {
                   roomId,
@@ -1674,7 +1651,7 @@ function useEditRoomComment(
   roomId: string
 ): (options: EditCommentOptions) => void {
   const client = useClient();
-  return React.useCallback(
+  return useCallback(
     ({ threadId, commentId, body, attachments }: EditCommentOptions): void => {
       const editedAt = new Date();
 
@@ -1701,7 +1678,7 @@ function useEditRoomComment(
         return;
       }
 
-      const optimisticUpdateId = store.addOptimisticUpdate({
+      const optimisticId = store.optimisticUpdates.add({
         type: "edit-comment",
         comment: {
           ...comment,
@@ -1718,12 +1695,12 @@ function useEditRoomComment(
         .then(
           (editedComment) => {
             // Replace the optimistic update by the real thing
-            store.editComment(threadId, optimisticUpdateId, editedComment);
+            store.editComment(threadId, optimisticId, editedComment);
           },
           (err: Error) =>
             onMutationFailure(
               err,
-              optimisticUpdateId,
+              optimisticId,
               (error) =>
                 new EditCommentError(error, {
                   roomId,
@@ -1756,13 +1733,13 @@ function useDeleteComment() {
 function useDeleteRoomComment(roomId: string) {
   const client = useClient();
 
-  return React.useCallback(
+  return useCallback(
     ({ threadId, commentId }: DeleteCommentOptions): void => {
       const deletedAt = new Date();
 
       const { store, onMutationFailure } = getRoomExtrasForClient(client);
 
-      const optimisticUpdateId = store.addOptimisticUpdate({
+      const optimisticId = store.optimisticUpdates.add({
         type: "delete-comment",
         threadId,
         commentId,
@@ -1775,17 +1752,12 @@ function useDeleteRoomComment(roomId: string) {
         .then(
           () => {
             // Replace the optimistic update by the real thing
-            store.deleteComment(
-              threadId,
-              optimisticUpdateId,
-              commentId,
-              deletedAt
-            );
+            store.deleteComment(threadId, optimisticId, commentId, deletedAt);
           },
           (err: Error) =>
             onMutationFailure(
               err,
-              optimisticUpdateId,
+              optimisticId,
               (error) =>
                 new DeleteCommentError(error, {
                   roomId,
@@ -1808,14 +1780,14 @@ function useAddReaction<M extends BaseMetadata>() {
  */
 function useAddRoomCommentReaction<M extends BaseMetadata>(roomId: string) {
   const client = useClient();
-  return React.useCallback(
+  return useCallback(
     ({ threadId, commentId, emoji }: CommentReactionOptions): void => {
       const createdAt = new Date();
       const userId = getCurrentUserId(client);
 
       const { store, onMutationFailure } = getRoomExtrasForClient<M>(client);
 
-      const optimisticUpdateId = store.addOptimisticUpdate({
+      const optimisticId = store.optimisticUpdates.add({
         type: "add-reaction",
         threadId,
         commentId,
@@ -1833,7 +1805,7 @@ function useAddRoomCommentReaction<M extends BaseMetadata>(roomId: string) {
             // Replace the optimistic update by the real thing
             store.addReaction(
               threadId,
-              optimisticUpdateId,
+              optimisticId,
               commentId,
               addedReaction,
               createdAt
@@ -1842,7 +1814,7 @@ function useAddRoomCommentReaction<M extends BaseMetadata>(roomId: string) {
           (err: Error) =>
             onMutationFailure(
               err,
-              optimisticUpdateId,
+              optimisticId,
               (error) =>
                 new AddReactionError(error, {
                   roomId,
@@ -1873,14 +1845,14 @@ function useRemoveReaction() {
  */
 function useRemoveRoomCommentReaction(roomId: string) {
   const client = useClient();
-  return React.useCallback(
+  return useCallback(
     ({ threadId, commentId, emoji }: CommentReactionOptions): void => {
       const userId = getCurrentUserId(client);
 
       const removedAt = new Date();
 
       const { store, onMutationFailure } = getRoomExtrasForClient(client);
-      const optimisticUpdateId = store.addOptimisticUpdate({
+      const optimisticId = store.optimisticUpdates.add({
         type: "remove-reaction",
         threadId,
         commentId,
@@ -1896,7 +1868,7 @@ function useRemoveRoomCommentReaction(roomId: string) {
             // Replace the optimistic update by the real thing
             store.removeReaction(
               threadId,
-              optimisticUpdateId,
+              optimisticId,
               commentId,
               emoji,
               userId,
@@ -1906,7 +1878,7 @@ function useRemoveRoomCommentReaction(roomId: string) {
           (err: Error) =>
             onMutationFailure(
               err,
-              optimisticUpdateId,
+              optimisticId,
               (error) =>
                 new RemoveReactionError(error, {
                   roomId,
@@ -1936,7 +1908,7 @@ function useMarkThreadAsRead() {
  */
 function useMarkRoomThreadAsRead(roomId: string) {
   const client = useClient();
-  return React.useCallback(
+  return useCallback(
     (threadId: string) => {
       const { store, onMutationFailure } = getRoomExtrasForClient(client);
       const inboxNotification = Object.values(
@@ -1951,7 +1923,7 @@ function useMarkRoomThreadAsRead(roomId: string) {
 
       const now = new Date();
 
-      const optimisticUpdateId = store.addOptimisticUpdate({
+      const optimisticId = store.optimisticUpdates.add({
         type: "mark-inbox-notification-as-read",
         inboxNotificationId: inboxNotification.id,
         readAt: now,
@@ -1965,16 +1937,16 @@ function useMarkRoomThreadAsRead(roomId: string) {
         .then(
           () => {
             // Replace the optimistic update by the real thing
-            store.updateInboxNotification(
+            store.markInboxNotificationRead(
               inboxNotification.id,
-              optimisticUpdateId,
-              (inboxNotification) => ({ ...inboxNotification, readAt: now })
+              now,
+              optimisticId
             );
           },
           (err: Error) => {
             onMutationFailure(
               err,
-              optimisticUpdateId,
+              optimisticId,
               (error) =>
                 new MarkInboxNotificationAsReadError(error, {
                   inboxNotificationId: inboxNotification.id,
@@ -2004,12 +1976,12 @@ function useMarkThreadAsResolved() {
  */
 function useMarkRoomThreadAsResolved(roomId: string) {
   const client = useClient();
-  return React.useCallback(
+  return useCallback(
     (threadId: string) => {
       const updatedAt = new Date();
 
       const { store, onMutationFailure } = getRoomExtrasForClient(client);
-      const optimisticUpdateId = store.addOptimisticUpdate({
+      const optimisticId = store.optimisticUpdates.add({
         type: "mark-thread-as-resolved",
         threadId,
         updatedAt,
@@ -2022,7 +1994,7 @@ function useMarkRoomThreadAsResolved(roomId: string) {
             // Replace the optimistic update by the real thing
             store.patchThread(
               threadId,
-              optimisticUpdateId,
+              optimisticId,
               { resolved: true },
               updatedAt
             );
@@ -2030,7 +2002,7 @@ function useMarkRoomThreadAsResolved(roomId: string) {
           (err: Error) =>
             onMutationFailure(
               err,
-              optimisticUpdateId,
+              optimisticId,
               (error) =>
                 new MarkThreadAsResolvedError(error, {
                   roomId,
@@ -2059,12 +2031,12 @@ function useMarkThreadAsUnresolved() {
  */
 function useMarkRoomThreadAsUnresolved(roomId: string) {
   const client = useClient();
-  return React.useCallback(
+  return useCallback(
     (threadId: string) => {
       const updatedAt = new Date();
 
       const { store, onMutationFailure } = getRoomExtrasForClient(client);
-      const optimisticUpdateId = store.addOptimisticUpdate({
+      const optimisticId = store.optimisticUpdates.add({
         type: "mark-thread-as-unresolved",
         threadId,
         updatedAt,
@@ -2077,7 +2049,7 @@ function useMarkRoomThreadAsUnresolved(roomId: string) {
             // Replace the optimistic update by the real thing
             store.patchThread(
               threadId,
-              optimisticUpdateId,
+              optimisticId,
               { resolved: false },
               updatedAt
             );
@@ -2085,7 +2057,7 @@ function useMarkRoomThreadAsUnresolved(roomId: string) {
           (err: Error) =>
             onMutationFailure(
               err,
-              optimisticUpdateId,
+              optimisticId,
               (error) =>
                 new MarkThreadAsUnresolvedError(error, {
                   roomId,
@@ -2110,7 +2082,7 @@ function useThreadSubscription(threadId: string): ThreadSubscription {
 
   const signal = store.outputs.threadifications;
 
-  const selector = React.useCallback(
+  const selector = useCallback(
     (state: SignalType<typeof signal>): ThreadSubscription => {
       const notification = state.sortedNotifications.find(
         (inboxNotification) =>
@@ -2153,7 +2125,7 @@ function useRoomNotificationSettings(): [
 
   const poller = getOrCreateNotificationsSettingsPollerForRoomId(room.id);
 
-  React.useEffect(
+  useEffect(
     () => {
       void store.waitUntilRoomNotificationSettingsLoaded(room.id);
     }
@@ -2167,7 +2139,7 @@ function useRoomNotificationSettings(): [
     //    *next* render after that, a *new* fetch/promise will get created.
   );
 
-  React.useEffect(() => {
+  useEffect(() => {
     poller.inc();
     poller.pollNowIfStale();
     return () => {
@@ -2175,19 +2147,19 @@ function useRoomNotificationSettings(): [
     };
   }, [poller]);
 
-  // TODO(vincent+nimesh) There is a disconnect between this getter and
-  // subscriber! It's unclear why the getNotificationSettingsLoadingState
-  // getter should be paired with subscribe2 and not subscribe1 from the
-  // outside! (The reason is that getNotificationSettingsLoadingState
-  // internally uses `get2` not `get1`.) This is strong evidence that
-  // getNotificationSettingsLoadingState itself wants to be a Signal! Once we
-  // make it a Signal, we can simply use `useSignal()` here! ❤️
-  const getter = React.useCallback(
+  // XXX_vincent There is a disconnect between this getter and subscriber! It's unclear
+  // why the getNotificationSettingsLoadingState getter should be paired with
+  // subscribe2 and not subscribe1 from the outside! (The reason is that
+  // getNotificationSettingsLoadingState internally uses `get2` not `get1`.)
+  // This is strong evidence that getNotificationSettingsLoadingState itself
+  // wants to be a Signal! Once we make it a Signal, we can simply use
+  // `useSignal()` here! ❤️
+  const getter = useCallback(
     () => store.getNotificationSettingsLoadingState(room.id),
     [store, room.id]
   );
 
-  // TODO(vincent+nimesh) Turn this into a useSignal
+  // XXX_vincent Turn this into a useSignal
   const settings = useSyncExternalStoreWithSelector(
     store.subscribe2,
     getter,
@@ -2196,7 +2168,7 @@ function useRoomNotificationSettings(): [
     shallow2
   );
 
-  return React.useMemo(() => {
+  return useMemo(() => {
     return [settings, updateRoomNotificationSettings];
   }, [settings, updateRoomNotificationSettings]);
 }
@@ -2226,7 +2198,7 @@ function useRoomNotificationSettingsSuspense(): [
   assert(!settings.error, "Did not expect error");
   assert(!settings.isLoading, "Did not expect loading");
 
-  return React.useMemo(() => {
+  return useMemo(() => {
     return [settings, updateRoomNotificationSettings];
   }, [settings, updateRoomNotificationSettings]);
 }
@@ -2240,11 +2212,11 @@ function useRoomNotificationSettingsSuspense(): [
 function useHistoryVersionData(
   versionId: string
 ): HistoryVersionDataAsyncResult {
-  const [state, setState] = React.useState<HistoryVersionDataAsyncResult>({
+  const [state, setState] = useState<HistoryVersionDataAsyncResult>({
     isLoading: true,
   });
   const room = useRoom();
-  React.useEffect(() => {
+  useEffect(() => {
     setState({ isLoading: true });
     const load = async () => {
       try {
@@ -2287,18 +2259,18 @@ function useHistoryVersions(): HistoryVersionsAsyncResult {
 
   const poller = getOrCreateVersionsPollerForRoomId(room.id);
 
-  React.useEffect(() => {
+  useEffect(() => {
     poller.inc();
     poller.pollNowIfStale();
     return () => poller.dec();
   }, [poller]);
 
-  const getter = React.useCallback(
+  const getter = useCallback(
     () => store.getRoomVersionsLoadingState(room.id),
     [store, room.id]
   );
 
-  React.useEffect(
+  useEffect(
     () => {
       void store.waitUntilRoomVersionsLoaded(room.id);
     }
@@ -2312,7 +2284,7 @@ function useHistoryVersions(): HistoryVersionsAsyncResult {
     //    *next* render after that, a *new* fetch/promise will get created.
   );
 
-  // TODO(vincent+nimesh) There is a disconnect between this getter and subscriber! It's unclear
+  // XXX_vincent There is a disconnect between this getter and subscriber! It's unclear
   // why the getRoomVersionsLoadingState getter should be paired with
   // subscribe3 and not subscribe1 from the outside! (The reason is that
   // getRoomVersionsLoadingState internally uses `get3` not `get1`.) This is
@@ -2359,10 +2331,10 @@ function useHistoryVersionsSuspense(): HistoryVersionsAsyncSuccess {
 function useUpdateRoomNotificationSettings() {
   const client = useClient();
   const room = useRoom();
-  return React.useCallback(
+  return useCallback(
     (settings: Partial<RoomNotificationSettings>) => {
       const { store, onMutationFailure } = getRoomExtrasForClient(client);
-      const optimisticUpdateId = store.addOptimisticUpdate({
+      const optimisticId = store.optimisticUpdates.add({
         type: "update-notification-settings",
         roomId: room.id,
         settings,
@@ -2371,16 +2343,12 @@ function useUpdateRoomNotificationSettings() {
       room.updateNotificationSettings(settings).then(
         (settings) => {
           // Replace the optimistic update by the real thing
-          store.updateRoomNotificationSettings_confirmOptimisticUpdate(
-            room.id,
-            optimisticUpdateId,
-            settings
-          );
+          store.updateRoomNotificationSettings(room.id, optimisticId, settings);
         },
         (err: Error) =>
           onMutationFailure(
             err,
-            optimisticUpdateId,
+            optimisticId,
             (error) =>
               new UpdateNotificationSettingsError(error, {
                 roomId: room.id,
@@ -2583,12 +2551,12 @@ function useRoomAttachmentUrl(
   const store =
     client[kInternal].httpClient.getOrCreateAttachmentUrlsStore(roomId);
 
-  const getAttachmentUrlState = React.useCallback(
+  const getAttachmentUrlState = useCallback(
     () => store.getState(attachmentId),
     [store, attachmentId]
   );
 
-  React.useEffect(() => {
+  useEffect(() => {
     // NOTE: .get() will trigger any actual fetches, whereas .getState() will not
     void store.get(attachmentId);
   }, [store, attachmentId]);
@@ -2612,7 +2580,7 @@ function useAttachmentUrlSuspense(attachmentId: string) {
   const room = useRoom();
   const { attachmentUrlsStore } = room[kInternal];
 
-  const getAttachmentUrlState = React.useCallback(
+  const getAttachmentUrlState = useCallback(
     () => attachmentUrlsStore.getState(attachmentId),
     [attachmentUrlsStore, attachmentId]
   );
@@ -2648,7 +2616,7 @@ function useRoomPermissions(roomId: string) {
   const client = useClient();
   const store = getRoomExtrasForClient(client).store;
   return useSignal(
-    store.permissionHintsByRoomId,
+    store.permissionHints.signal,
     (hints) => hints[roomId] ?? new Set()
   );
 }
@@ -2840,7 +2808,7 @@ const _useEventListener: TypedBundle["useEventListener"] = useEventListener;
 
 /**
  * Returns the presence of the current user of the current room, and a function to update it.
- * It is different from the setState function returned by the useState hook from React.
+ * It is different from the setState function returned by the useState hook from
  * You don't need to pass the full presence object to update it.
  *
  * @example